+++ src/com/android/launcher3/dragndrop/DragView.java
@@ -66,8 +66,6 @@ import com.android.launcher3.util.Themes;
 import com.android.launcher3.util.Thunk;
 import com.android.launcher3.widget.PendingAddShortcutInfo;
 
-import org.zimmob.zimlx.pixelify.ClockIconDrawable;
-
 import java.util.Arrays;
 import java.util.List;
 
@@ -323,8 +321,6 @@ public class DragView extends View {
                     .resolveActivity(info.getIntent(), info.user);
             outObj[0] = activityInfo;
             if (activityInfo != null) {
-                if (Utilities.isAnimatedClock(getContext(), activityInfo.getComponentName()))
-                    return ClockIconDrawable.Companion.create(getContext());
                 return appState.getIconCache().getFullResIcon(activityInfo, false);
             }
             return (activityInfo != null) ? appState.getIconCache()