+++ src/org/zimmob/zimlx/ConfigFlags.java
@@ -1,9 +1,0 @@
-package org.zimmob.zimlx;
-
-public class ConfigFlags {
-    //Folder Shape
-    public static final int FOLDER_SHAPE_CIRCLE = 0;
-    public static final int FOLDER_SHAPE_CIRCLE_SHADOW = 1;
-    public static final int FOLDER_SHAPE_SQUARE = 2;
-    public static final int FOLDER_SHAPE_SQUARE_SHADOW = 3;
-}