+++ src/org/zimmob/zimlx/minibar/Minibar.java
@@ -26,6 +26,7 @@ public class Minibar {
             new ActionDisplayItem(Action.VolumeDialog, "VolumeDialog", Launcher.getLauncher(mContext).getString(R.string.minibar_7), R.drawable.ic_volume_up_black_24dp, 71),
             new ActionDisplayItem(Action.DeviceSettings, "DeviceSettings", Launcher.getLauncher(mContext).getString(R.string.minibar_4), R.drawable.ic_build, 25),
             new ActionDisplayItem(Action.MobileNetworkSettings, "MobileNetworkSettings", Launcher.getLauncher(mContext).getString(R.string.minibar_10), R.drawable.ic_network_24dp, 46),
+            new ActionDisplayItem(Action.AppSettings, "AppSettings", Launcher.getLauncher(mContext).getString(R.string.minibar_11), R.drawable.ic_font_download, 54),
             new ActionDisplayItem(Action.AppDrawer, "AppDrawer", Launcher.getLauncher(mContext).getString(R.string.minibar_8), R.drawable.ic_apps_dark_24dp, 73)
     };
 
@@ -40,6 +41,10 @@ public class Minibar {
                 break;
             case MobileNetworkSettings: {
                 context.startActivity(new Intent(Settings.ACTION_DATA_ROAMING_SETTINGS));
+                break;
+            }
+            case AppSettings: {
+                context.startActivity(new Intent(Settings.ACTION_MANAGE_ALL_APPLICATIONS_SETTINGS));
                 break;
             }
             case SetWallpaper:
@@ -106,7 +111,7 @@ public class Minibar {
     }
 
     public enum Action {
-        EditMinibar, SetWallpaper, LockScreen, DeviceSettings, LauncherSettings, VolumeDialog, AppDrawer, MobileNetworkSettings,
+        EditMinibar, SetWallpaper, LockScreen, DeviceSettings, LauncherSettings, VolumeDialog, AppDrawer, MobileNetworkSettings, AppSettings
     }
 
     public static class ActionItem {