+++ src/org/zimmob/zimlx/util/ZimFlags.java
@@ -1,0 +1,20 @@
+package org.zimmob.zimlx.util;
+
+public class ZimFlags {
+
+    public static String ACCENT_COLOR = "pref_key__accent_color";
+    public static String MINIBAR_COLOR = "pref_key__minibar_color";
+    public static String PRIMARY_COLOR = "pref_key__primary_color";
+    public static String ICON_PACK = "pref_icon_pack";
+    public static String DESKTOP_ICON_SCALE = "pref_iconScaleSB";
+    public static String HOTSEAT_HEIGHT_SCALE = "pref_hotseatHeightScale";
+    public static String USE_PIXEL_ICONS = "pref_key__pixel_icons";
+    public static String OVERRIDE_LAUNCHER_THEME = "pref_override_launcher_theme";
+    public static String APPDRAWER_ICON_SCALE = "pref_allAppsIconScale";
+
+    //Folder Shape
+    public static final int FOLDER_SHAPE_CIRCLE = 0;
+    public static final int FOLDER_SHAPE_CIRCLE_SHADOW = 1;
+    public static final int FOLDER_SHAPE_SQUARE = 2;
+    public static final int FOLDER_SHAPE_SQUARE_SHADOW = 3;
+}